Add files via upload
This is a Forecasting Project for my Master of Analytics program at RMIT. This report consists of two parts. The first part aims to analyse and forecast the amount of horizontal solar radiation reaching the ground at a particular location over the globe. To achieve this task, I applied time series regression method to fit distributed lag models with the monthly precipitation series as an independent explanatory series. Moreover, I will also be using exponential smoothing methods. The expectation is to find the best model to provide an accurate forecast of the amount of horizontal solar radiation.

The second part of the report analyses the correlation between the quarterly Residential Property Price Index (PPI) in Melbourne and quarterly population change over the previous quarter in Victoria between September 2003 and December 2016. The data will be explored and investigated to demonstrate whether the correlation between these two series is spurious or not.

Models utilised:
- Finite distributed lag model
- Polynomial distributed lag model
- Koyck model
- Dynamic lag model
- Autoregressive distributed lag model
- Exponential smoothing
- State-space model