Add files via upload
This is a data visualisation project that I created as part of my studies at RMIT. I am a huge NBA fan so I chose to make this visualisation project as an analysis of the different methods of scoring employed by the top 20 highest scorers of the NBA’s 2022/2023 season to determine what these players have in common and what makes them such effective scorers. Furthermore, the project aims to discover whether the volume and/or efficiency of a type of shot can set the most elite scorers in the NBA apart from the rest.​

The dataset used for the presentation was published by Basketball Reference. This dataset includes the per-game data of the leading scorers in the NBA who played more than 57 games, which is 70% of the available games in a season. This ensures that the data accurately reflects the scoring ability of players throughout a relatively equal sample size of games. ​