Add files via upload
This project is a part of my Master of Analytics program at RMIT. I worked in a group with three members for this project, the other two members are credited in the report. This report employed the Obesity Levels dataset from Kaggle (Mehrparvar, 2024). This dataset includes information about the estimation of obesity levels in Mexico, Peru, and Colombia using the individuals' physical condition and eating habits.

The report employs many different visualisation techniques to explore and find trends in each variable of the data. The project was made using Python, more specifically, Jupyter Notebook. Pandas and NumPy are also utilised in this project.