Add files via upload
In this project, I created a database about Covid-19 vaccination statistics from different countries across the globe. The data was taken from: "https://github.com/owid/covid-19-data/tree/master/public/data/vaccinations". The database is then used to complete a series of tasks and retrieve useful information about Covid-19 vaccinations. The folder contains the CSV files that were used to create the database, the steps of the design process, the actual database, and the SQLs that were used to retrieve information.